Refactor code and fix some bugs Add the user id in the single user end point in the ReadME.md Seperate endpoints to handle authentication from those to handle recipes and categories by putting them in auth.py Fix the bug of recurring errors when server is not restatrted The bug was fixed in validator.py Change includes in api/__init__.py to include auth
